Fix warnings revealed by the test

+++ sycl/include/CL/sycl/detail/cg.hpp
@@ -208,7 +208,7 @@ public:
     // If local size for host is not set explicitly, let's adjust it to 1,
     // so nd_range_error for zero local size is not thrown.
     if (AdjustedRange.LocalSize[0] == 0)
-      for (int I = 0; I < AdjustedRange.Dims; ++I)
+      for (size_t I = 0; I < AdjustedRange.Dims; ++I)
         AdjustedRange.LocalSize[I] = 1;
     if (HPI)
       HPI->start();
